% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire}
\format{
A data frame with 9794 rows and 10 variables:
\describe{
\item{\code{Location}}{integer COLUMN_DESCRIPTION}
\item{\code{Tag}}{integer COLUMN_DESCRIPTION}
\item{\code{Date}}{character COLUMN_DESCRIPTION}
\item{\code{Entry}}{character COLUMN_DESCRIPTION}
\item{\code{Exit}}{character COLUMN_DESCRIPTION}
\item{\code{Ent Wt}}{double COLUMN_DESCRIPTION}
\item{\code{Ext Wt}}{double COLUMN_DESCRIPTION}
\item{\code{Consumed}}{double COLUMN_DESCRIPTION}
\item{\code{Weight}}{double COLUMN_DESCRIPTION}
\item{\code{Topup Amount}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
fire
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
