% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintyr_example.R
\name{mintyr_example}
\alias{mintyr_example}
\title{Get path to mintyr examples}
\usage{
mintyr_example(path)
}
\arguments{
\item{path}{Name of file.}
}
\description{
mintyr comes bundled with a number of sample files in
its \code{inst/extdata} directory. Use \code{mintyr_example()} to retrieve the path to one
example.
}
\examples{
mintyr_example("csv_test1.csv")
}
\seealso{
\code{\link[=mintyr_examples]{mintyr_examples()}}
}
