% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintyr_examples.R
\name{mintyr_examples}
\alias{mintyr_examples}
\title{Get path to one example}
\usage{
mintyr_examples(pattern = NULL)
}
\arguments{
\item{pattern}{A regular expression of filenames to match. If \code{NULL}, all available files are returned.}
}
\description{
vroom comes bundled with a number of sample files in
its 'inst/extdata' directory. Use \code{mintyr_examples()} to list all the
available examples.
}
\examples{
mintyr_examples()
}
\seealso{
\code{\link[=mintyr_example]{mintyr_example()}}
}
