% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nedap}
\alias{nedap}
\title{nedap}
\format{
A data frame with 31863 rows and 9 variables:
\describe{
\item{\code{animal_number}}{integer COLUMN_DESCRIPTION}
\item{\code{lifenumber}}{logical COLUMN_DESCRIPTION}
\item{\code{responder}}{integer COLUMN_DESCRIPTION}
\item{\code{location}}{integer COLUMN_DESCRIPTION}
\item{\code{visit_time}}{double COLUMN_DESCRIPTION}
\item{\code{duration}}{integer COLUMN_DESCRIPTION}
\item{\code{state}}{integer COLUMN_DESCRIPTION}
\item{\code{weight}}{integer COLUMN_DESCRIPTION}
\item{\code{feed_intake}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
nedap
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
