% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_export.R
\name{nest_export}
\alias{nest_export}
\title{Export Nested Data Frames to Files Based on Grouping Columns}
\usage{
nest_export(
  nest_dt,
  group_cols = NULL,
  nest_col = NULL,
  export_path = tempdir(),
  file_type = "txt"
)
}
\arguments{
\item{nest_dt}{A \code{data.frame} or \code{data.table} containing at least one nested column of data frames or data tables. This is the input data from which nested data will be exported.}

\item{group_cols}{A non-empty character vector specifying the columns in \code{nest_dt} to use for grouping when exporting. Directories will be created based on combinations of these groupings.}

\item{nest_col}{A single character string specifying the name of the column in \code{nest_dt} that contains the nested data frames or data tables to be exported. This name also serves as the base filename for the exported files.}

\item{export_path}{A single character string specifying the base directory path where the files will be exported. If the directory does not exist, it will be created. Defaults to \code{tempdir()}.}

\item{file_type}{A character string specifying the type of file to export: either \code{"txt"} for tab-separated values or \code{"csv"} for comma-separated values. Defaults to \code{"txt"}. Case-insensitive.}
}
\value{
An integer representing the total number of files exported successfully.
}
\description{
The \code{nest_export} function exports nested data frames or data tables stored within a column of a data table (\code{nest_dt}) to files. It expands the nested data, groups it by specified columns (\code{group_cols}), and writes the data to files within directories reflecting the groupings. The files can be saved as tab-separated (\code{.txt}) or comma-separated (\code{.csv}) values.
}
\details{
The function performs the following steps:
\enumerate{
\item Validates the input parameters, ensuring that \code{nest_col} and \code{group_cols} are provided and exist within \code{nest_dt}. It checks that \code{file_type} is either \code{"txt"} or \code{"csv"}.
\item Identifies the nested column specified by \code{nest_col} and expands it by unnesting the data frames or data tables it contains.
\item Groups the expanded data by the columns specified in \code{group_cols}.
\item For each group, constructs a directory path based on the grouping values under \code{export_path} and writes the data to a file named after \code{nest_col}, using the specified \code{file_type}.
\item Utilizes multiple threads when writing files for efficiency, using \code{data.table::fwrite}.
}
}
\note{
\itemize{
\item The \code{nest_dt} must contain at least one nested column of data frames or data tables.
\item The \code{group_cols} must be a non-empty character vector corresponding to columns in \code{nest_dt}.
\item The \code{nest_col} must be a single character string corresponding to a column name in \code{nest_dt}.
\item If \code{export_path} does not exist, the function will create it.
\item The function leverages multi-threading via \code{data.table::fwrite} for faster file writing.
\item The original \code{nest_dt} is not modified; a modified copy is used internally.
\item If an error occurs during the export process, the function stops and returns an error message.
}
}
\examples{
# Data Nesting Operation
dt_nest <- w2l_nest(data = iris, cols2l = 1:2, by = "Species")
# Export Nested Data
nest_export(nest_dt = dt_nest, nest_col = "data",
            group_cols = c("name", "Species"))
# Check Export Results
list.files(path = tempdir(), pattern = "txt", recursive = TRUE)
}
