% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire}
\format{
A data frame with 9794 rows and 10 variables:
\describe{
\item{\code{Location}}{integer Feeding station identification number}
\item{\code{Tag}}{integer Animal electronic tag number}
\item{\code{Date}}{character Date of feeding visit}
\item{\code{Entry}}{character Time when animal entered feeding station}
\item{\code{Exit}}{character Time when animal left feeding station}
\item{\code{Ent Wt}}{double Feed weight at entry (kg)}
\item{\code{Ext Wt}}{double Feed weight at exit (kg)}
\item{\code{Consumed}}{double Amount of feed consumed (kg)}
\item{\code{Weight}}{double Animal body weight (kg)}
\item{\code{Topup Amount}}{double Amount of feed added to bin (kg)}
}
}
\usage{
fire
}
\description{
Feeding behavior dataset from Fire system
}
\keyword{datasets}
