% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filename.R
\name{get_filename}
\alias{get_filename}
\title{Extract specific parts of the file path}
\usage{
get_filename(paths, remove_extension = TRUE, remove_path = TRUE)
}
\arguments{
\item{paths}{character vector. A string or a character vector representing the full path(s) to the file(s).}

\item{remove_extension}{logical, optional. If TRUE, the file extensions are removed from the filenames. Default is TRUE.}

\item{remove_path}{logical, optional. If TRUE, the directory paths are removed, leaving only the filenames. Default is TRUE.}
}
\value{
character vector. A string or a character vector representing the processed file name(s).
}
\description{
This function extracts the base name of the file(s) from the given path(s), with options to remove the file extension and/or the directory path.
}
\note{
Ensure that the \code{paths} parameter is a character vector. The function will stop with an error if \code{paths} is not a character vector.
}
\examples{
xlsx_files <- mintyr_example(mintyr_examples("xlsx_test"))
xlsx_files
get_filename(xlsx_files)
get_filename(xlsx_files, remove_extension = FALSE)
get_filename(xlsx_files, remove_path = FALSE)
}
