% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\title{Import CSV Files Using Specified Package}
\usage{
import_csv(file, package = "data.table", rbind = TRUE, rbind_label = NULL, ...)
}
\arguments{
\item{file}{character vector of file paths. The paths to the CSV files to be imported.. These paths must point to existing files.}

\item{package}{character, optional. Specifies which package to use for reading the CSV files. Default is 'data.table'. Valid options are \link{data.table}, \link{vroom}, and \link{arrow}.}

\item{rbind}{logical, optional. If TRUE and more than one file is specified, it binds rows from different files into one dataset. Default is TRUE. Note: For \code{vroom}, row binding is inherent and cannot be turned off.}

\item{rbind_label}{character, optional. The name of the column to add to the dataset that indicates the file origin. This is useful when row binding is used. Default is NULL.}

\item{...}{Additional arguments passed to \code{\link[=import_csv]{import_csv()}}.The arguments vary depending on the R package chosen.}
}
\value{
Depending on the input and settings, returns a data.table or a list of data.tables.
If \code{rbind} is TRUE and more than one file is processed, it returns a single data.table with all files combined.
If \code{rbind} is FALSE or only one file is provided, it returns the data table from the single file processed.
}
\description{
The \code{import_csv} function imports one or multiple CSV files using a specified package (\code{data.table}, \code{vroom}, or \code{arrow}).
It supports row binding of multiple files into a single dataset with optional labeling of the origin file.
}
\note{
\itemize{
\item The function stops and throws an error if any file does not exist or if an unsupported package is specified.
\item The \code{vroom} package inherently binds rows when reading multiple files and does not support turning off this feature.
}
}
\examples{
csv_files <- mintyr_example(mintyr_examples("csv_test"))
csv_files
import_csv(csv_files)
}
\seealso{
\code{\link[data.table:fread]{data.table::fread()}}

\code{\link[vroom:vroom]{vroom::vroom()}}

\code{\link[arrow:read_delim_arrow]{arrow::read_csv_arrow()}}
}
