% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xlsx.R
\name{import_xlsx}
\alias{import_xlsx}
\title{Import list of xlsx sheets}
\usage{
import_xlsx(file, rbind = TRUE, sheet = NULL, ...)
}
\arguments{
\item{file}{character vector of file paths. Paths to the Excel files to be imported. These paths must point to existing files.}

\item{rbind}{logical, optional. If TRUE and more than one file is specified, it binds rows from different files into one dataset. Default is TRUE. Note: For \code{vroom}, row binding is inherent and cannot be turned off.}

\item{sheet}{numeric vector or NULL, optional. Specifies the sheets to be read from each file. If NULL, all sheets are read. If numeric, only the specified sheets (by index) are read. Default is NULL.}

\item{...}{Additional arguments to be passed to the \code{\link[readxl:read_excel]{readxl::read_excel()}} function, such as col_types to specify column data types.}
}
\value{
Depending on the \code{rbind} parameter:
\itemize{
\item If \code{rbind} is TRUE, returns a single \code{data.table} combining all files and sheets, with an additional 'excel_name' column indicating the source file.
\item If \code{rbind} is FALSE, returns a named list of \code{data.tables}, each representing data from a sheet, named by combining the file and sheet names.
}
}
\description{
The \code{import_xlsx} function reads data from Excel files, optionally from specified sheets, and can either merge data from multiple files or keep them separate.
}
\note{
Make sure that the \code{file} paths provided are valid and exist. The function will stop with an error if any file does not exist or if the sheet indices are out of range.
}
\examples{
xlsx_files <- mintyr_example(mintyr_examples("xlsx_test"))
xlsx_files
import_xlsx(xlsx_files)
}
