% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nedap}
\alias{nedap}
\title{nedap}
\format{
A data frame with 31863 rows and 9 variables:
\describe{
\item{\code{animal_number}}{integer Animal identification number}
\item{\code{lifenumber}}{logical Life number of the animal}
\item{\code{responder}}{integer Responder identification number}
\item{\code{location}}{integer Feeding station location}
\item{\code{visit_time}}{double Time of feeding visit}
\item{\code{duration}}{integer Duration of feeding visit (minutes)}
\item{\code{state}}{integer Status code}
\item{\code{weight}}{integer Body weight (kg)}
\item{\code{feed_intake}}{integer Feed intake amount (kg)}
}
}
\usage{
nedap
}
\description{
Dairy cow feeding behavior dataset
}
\keyword{datasets}
