% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_cv.R
\name{nest_cv}
\alias{nest_cv}
\title{Apply Cross-Validation to Nested Data Tables}
\usage{
nest_cv(
  nest_dt,
  v = 10,
  repeats = 1,
  strata = NULL,
  breaks = 4,
  pool = 0.1,
  ...
)
}
\arguments{
\item{nest_dt}{data.table. The input data table containing nested data tables.}

\item{v}{The number of partitions of the data set.}

\item{repeats}{The number of times to repeat the V-fold partitioning.}

\item{strata}{A variable in \code{data} (single character or name) used to conduct
stratified sampling. When not \code{NULL}, each resample is created within the
stratification variable. Numeric \code{strata} are binned into quartiles.}

\item{breaks}{A single number giving the number of bins desired to stratify a
numeric stratification variable.}

\item{pool}{A proportion of data used to determine if a particular group is
too small and should be pooled into another group. We do not recommend
decreasing this argument below its default of 0.1 because of the dangers
of stratifying groups that are too small.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
data.table. A data.table with cross-validation splits applied to the nested data tables.
}
\description{
The \code{nest_cv} function performs cross-validation on nested data tables within a \code{data.table}. It supports various cross-validation configurations and handles nested lists of data tables.
}
\note{
\itemize{
\item The \code{nest_dt} parameter must be a \code{data.table} containing at least one nested list column of data tables.
\item The function will copy the input \code{data.table} to avoid modifying the original data.
}
}
\examples{
dt_nest <- w2l_nest(data = iris, cols2l = 1:2, by = "Species")
nest_cv(nest_dt = dt_nest, v = 2, repeats = 2)
}
