% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2l_nest.R
\name{w2l_nest}
\alias{w2l_nest}
\title{Reshape Wide Data to Long and Nest by Specified Columns}
\usage{
w2l_nest(data, cols2l, by = NULL)
}
\arguments{
\item{data}{data.frame or data.table. The input data in wide format.}

\item{cols2l}{numeric or character vector. Specifies the columns to reshape from wide to long format. Can be either numeric indices or column names.}

\item{by}{character vector, optional. Specifies the columns to group by. Default is NULL.}
}
\value{
data.table. A data.table with nested data in long format, grouped by specified columns if provided.
\itemize{
\item If \code{by} is NULL, returns a data.table nested by 'name'.
\item If \code{by} is specified, returns a data.table nested by 'name' and the grouping variables.
}
}
\description{
The \code{w2l_nest} function reshapes wide-format data to long-format and nests it by specified columns.
It is designed to handle both data.frame and data.table objects and provides options for grouping the data.
}
\note{
\itemize{
\item The \code{cols2l} parameter should be either numeric indices or a character vector of column names.
\item Ensure all grouping variables specified in \code{by} are present in the data.
\item The function converts \code{data.frame} to \code{data.table} if necessary.
}
}
\examples{
w2l_nest(data = iris, cols2l = 1:4)
w2l_nest(data = iris, cols2l = c("Sepal.Length", "Sepal.Width", "Petal.Length"), by = "Species")
}
