% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2l_split.R
\name{w2l_split}
\alias{w2l_split}
\title{Convert Wide Data to Long Format and Split into List of Data Tables}
\usage{
w2l_split(data, cols2l)
}
\arguments{
\item{data}{data.frame or data.table. The input data in wide format.}

\item{cols2l}{numeric or character vector. Specifies the columns to reshape from wide to long format. Can be either numeric indices or column names.}
}
\value{
A list of data tables. Each data table in the list corresponds to one of the specified columns, reshaped into long format.
}
\description{
The \code{w2l_split} function reshapes wide-format data to long-format and splits it into a list of data tables,
each corresponding to one of the specified columns.
}
\note{
\itemize{
\item The \code{cols2l} parameter should be either numeric indices or a character vector of column names.
\item Ensure all specified columns in \code{cols2l} are present in the data.
\item The function converts \code{data.frame} to \code{data.table} if necessary.
}
}
\examples{
w2l_split(data = iris, cols2l = 1:3)
w2l_split(data = iris, cols2l = c("Sepal.Length", "Sepal.Width"))
}
